requirements.txt
openai
python_dotenv
spacy